refactor GrpcWebEndToEnd test data
