+++ test/RemoteMvvmTool.Tests/GrpcWebEndToEndTests.cs
@@ -139,23 +139,15 @@ public class GrpcWebEndToEndTests
         
         // Copy pre-built test files from the TestData directory (excluding the manual JS stub)
         var testDataDir = Path.Combine(rootDir, "test", "RemoteMvvmTool.Tests", "TestData", "GrpcWebEndToEnd");
-        if (Directory.Exists(testDataDir))
+        if (!Directory.Exists(testDataDir))
+            throw new DirectoryNotFoundException($"Test data directory not found at {testDataDir}");
+
+        foreach (var file in Directory.GetFiles(testDataDir))
         {
-            Console.WriteLine($"Copying test data from: {testDataDir}");
-            // Copy files but exclude the manual testviewmodelservice_pb.js stub
-            foreach (var file in Directory.GetFiles(testDataDir))
-            {
-                var fileName = Path.GetFileName(file);
-                if (fileName != "testviewmodelservice_pb.js") // Skip the manual stub
-                {
-                    File.Copy(file, Path.Combine(testProjectDir, fileName), true);
-                }
-            }
-        }
-        else
-        {
-            Console.WriteLine($"Test data directory not found at {testDataDir}, creating files dynamically");
-            CreateTestFiles(testProjectDir);
+            var fileName = Path.GetFileName(file);
+            if (fileName == "testviewmodelservice_pb.js") // Skip the manual stub, copied only if needed later
+                continue;
+            File.Copy(file, Path.Combine(testProjectDir, fileName), true);
         }
 
         // Copy node_modules from ThermalTest if available
@@ -174,417 +166,13 @@ public class GrpcWebEndToEndTests
         }
     }
 
-    static void CreateTestFiles(string testProjectDir)
+    static void CreateProtobufStub(string outDir)
     {
-        // Create package.json
-        var packageJson = @"{
-  ""name"": ""grpc-web-test"",
-  ""version"": ""1.0.0"",
-  ""dependencies"": {
-    ""grpc-web"": ""^1.4.2"",
-    ""google-protobuf"": ""^3.21.2"",
-    ""@types/google-protobuf"": ""^3.15.5""
-  }
-}";
-        File.WriteAllText(Path.Combine(testProjectDir, "package.json"), packageJson);
-
-        // Create tsconfig.json
-        var tsconfig = @"{
-  ""compilerOptions"": {
-    ""target"": ""es2020"",
-    ""module"": ""commonjs"",
-    ""strict"": false,
-    ""esModuleInterop"": true,
-    ""lib"": [""es2020"", ""dom""],
-    ""outDir"": ""dist"",
-    ""allowJs"": true,
-    ""moduleResolution"": ""node"",
-    ""types"": [""node""]
-  },
-  ""include"": [""**/*.ts"", ""**/*.js""],
-  ""exclude"": [""node_modules"", ""dist""]
-}";
-        File.WriteAllText(Path.Combine(testProjectDir, "tsconfig.json"), tsconfig);
-
-        // Create the main test TypeScript file
-        CreateTestTypeScriptFile(testProjectDir);
-        
-        // Create the protobuf JS stub
-        CreateProtobufStub(testProjectDir);
-    }
-
-    static void CreateTestTypeScriptFile(string testProjectDir)
-    {
-        var testTsContent = @"declare var process: any;
-
-(async () => {
-  console.log('Starting gRPC-Web test with generated protobuf parsing...');
-  
-  try {
-    // Try to require the generated protobuf files
-    let TestViewModelState: any;
-    let ThermalZoneComponentViewModelState: any;
-    
-    try {
-      // Import generated protobuf classes
-      const pb = require('./testviewmodelservice_pb.js');
-      TestViewModelState = pb.TestViewModelState;
-      ThermalZoneComponentViewModelState = pb.ThermalZoneComponentViewModelState;
-      console.log('Successfully loaded generated protobuf classes');
-    } catch (importError: any) {
-      console.warn('Could not load generated protobuf classes, using fallback parsing:', importError.message);
-      // Fall back to manual parsing if protobuf classes not available
-    }
-    
-    console.log('Making gRPC-web call...');
-    
-    const port = process.argv[2] || '5000';
-    const response = await fetch(`http://localhost:${port}/test_protos.TestViewModelService/GetState`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/grpc-web+proto' },
-      body: new Uint8Array([0,0,0,0,0]) // Empty protobuf message for GetState
-    });
-    
-    if (!response.ok) {
-      throw new Error(`HTTP ${response.status}: ${response.statusText}`);
-    }
-    
-    const buffer = await response.arrayBuffer();
-    const bytes = new Uint8Array(buffer);
-    
-    console.log('Response received:', bytes.length, 'bytes');
-    console.log('First 20 bytes:', Array.from(bytes.slice(0, 20)));
-    
-    if (bytes.length < 5) {
-      throw new Error('Response too short');
-    }
-    
-    // Parse gRPC-web response format: [compressed_flag][message_length][message_data]
-    const messageLength = (bytes[1] << 24) | (bytes[2] << 16) | (bytes[3] << 8) | bytes[4];
-    console.log('Message length from header:', messageLength);
-    
-    if (messageLength === 0) {
-      console.error('Server returned empty message - this suggests a server-side issue');
-      console.log('Full response bytes:', Array.from(bytes.slice(0, 100)));
-      throw new Error('Empty message returned from server');
-    }
-    
-    const messageBytes = bytes.slice(5, 5 + messageLength);
-    console.log('Protobuf message bytes:', messageBytes.length, 'bytes');
-    console.log('Message data:', Array.from(messageBytes.slice(0, 50)));
-    
-    let zones: any[] = [];
-    
-    if (TestViewModelState && ThermalZoneComponentViewModelState) {
-      try {
-        // Use generated protobuf classes to deserialize
-        const state = TestViewModelState.deserializeBinary(messageBytes);
-        console.log('Deserialized state using protobuf:', state);
-        
-        const zoneList = state.getZoneListList ? state.getZoneListList() : [];
-        console.log('Zone list from protobuf:', zoneList.length, 'items');
-        
-        zones = zoneList.map((zone: any, index: number) => ({
-          zone: zone.getZone ? zone.getZone() : index,
-          temperature: zone.getTemperature ? zone.getTemperature() : 0
-        }));
-      } catch (pbError: any) {
-        console.warn('Protobuf deserialization failed:', pbError.message);
-        console.log('Falling back to manual parsing...');
-        zones = parseZonesFromProtobuf(messageBytes);
-      }
-    } else {
-      console.log('Using manual protobuf parsing...');
-      zones = parseZonesFromProtobuf(messageBytes);
-    }
-    
-    console.log('Final parsed zones:', zones);
-    
-    if (zones.length < 2) {
-      throw new Error(`Expected at least 2 zones, got ${zones.length}`);
-    }
-    
-    if (zones[0].temperature !== 42 || zones[1].temperature !== 43) {
-      throw new Error(`Expected temperatures [42, 43], got [${zones[0].temperature}, ${zones[1].temperature}]`);
-    }
-    
-    console.log('? Test passed! Successfully retrieved collection from server using protobuf parsing');
-    console.log(`Zone 1: Zone=${zones[0].zone}, Temperature=${zones[0].temperature}`);
-    console.log(`Zone 2: Zone=${zones[1].zone}, Temperature=${zones[1].temperature}`);
-    
-  } catch (error: any) {
-    console.error('? Test failed:', error);
-    throw error;
-  }
-})().catch(e => { 
-  console.error('Unhandled error:', e); 
-  process.exit(1); 
-});
-
-// Fallback manual parsing functions
-function readVarint(bytes: Uint8Array, pos: number): { value: number; newPos: number } | null {
-  if (pos >= bytes.length) return null;
-  
-  let value = 0;
-  let shift = 0;
-  let currentPos = pos;
-  
-  while (currentPos < bytes.length && shift < 64) {
-    const byte = bytes[currentPos++];
-    value |= (byte & 0x7F) << shift;
-    
-    if ((byte & 0x80) === 0) {
-      return { value, newPos: currentPos };
-    }
-    
-    shift += 7;
-  }
-  
-  return null;
-}
-
-function parseZonesFromProtobuf(bytes: Uint8Array): any[] {
-  const zones: any[] = [];
-  let pos = 0;
-  
-  console.log('Manual parsing protobuf message of', bytes.length, 'bytes');
-  
-  while (pos < bytes.length) {
-    const tagResult = readVarint(bytes, pos);
-    if (!tagResult) break;
-    
-    const tag = tagResult.value;
-    pos = tagResult.newPos;
-    
-    const fieldNumber = tag >> 3;
-    const wireType = tag & 0x7;
-    
-    console.log(`Field ${fieldNumber}, wire type ${wireType}, pos ${pos}`);
-    
-    if (fieldNumber === 1 && wireType === 2) { // repeated zone_list field  
-      const lengthResult = readVarint(bytes, pos);
-      if (!lengthResult) break;
-      
-      const length = lengthResult.value;
-      pos = lengthResult.newPos;
-      
-      console.log(`Reading zone of length ${length} at pos ${pos}`);
-      
-      if (pos + length <= bytes.length) {
-        const zoneBytes = bytes.slice(pos, pos + length);
-        const zone = parseZone(zoneBytes);
-        zones.push(zone);
-        pos += length;
-        console.log('Parsed zone:', zone);
-      } else {
-        console.warn('Zone data exceeds buffer');
-        break;
-      }
-    } else {
-      console.log(`Skipping unknown field ${fieldNumber}, wire type ${wireType}`);
-      // Skip unknown field by reading the next varint (if wire type 0) or length-delimited data (if wire type 2)
-      if (wireType === 0) {
-        const skipResult = readVarint(bytes, pos);
-        if (skipResult) pos = skipResult.newPos;
-        else break;
-      } else if (wireType === 2) {
-        const lenResult = readVarint(bytes, pos);
-        if (lenResult) {
-          pos = lenResult.newPos + lenResult.value;
-        } else break;
-      } else {
-        break; // Unknown wire type
-      }
-    }
-  }
-  
-  return zones;
-}
-
-function parseZone(bytes: Uint8Array): any {
-  const zone: any = { zone: 0, temperature: 0 };
-  let pos = 0;
-  
-  console.log('Parsing zone from', bytes.length, 'bytes:', Array.from(bytes));
-  
-  while (pos < bytes.length) {
-    const tagResult = readVarint(bytes, pos);
-    if (!tagResult) break;
-    
-    const tag = tagResult.value;
-    pos = tagResult.newPos;
-    
-    const fieldNumber = tag >> 3;
-    console.log(`Zone field ${fieldNumber} at pos ${pos}`);
-    
-    if (fieldNumber === 1) { // zone field
-      const valueResult = readVarint(bytes, pos);
-      if (!valueResult) break;
-      zone.zone = valueResult.value;
-      pos = valueResult.newPos;
-      console.log('Zone value:', zone.zone);
-    } else if (fieldNumber === 2) { // temperature field  
-      const valueResult = readVarint(bytes, pos);
-      if (!valueResult) break;
-      zone.temperature = valueResult.value;
-      pos = valueResult.newPos;
-      console.log('Temperature value:', zone.temperature);
-    } else {
-      console.log('Unknown zone field:', fieldNumber);
-      // Skip unknown field
-      const skipResult = readVarint(bytes, pos);
-      if (skipResult) pos = skipResult.newPos;
-      else break;
-    }
-  }
-  
-  return zone;
-}";
-
-        File.WriteAllText(Path.Combine(testProjectDir, "test.ts"), testTsContent);
-    }
-
-    static void CreateProtobufStub(string testProjectDir)
-    {
-        var protobufStubContent = @"// Generated protobuf JavaScript stubs for testing
-exports.TestViewModelState = class {
-    constructor() { 
-        this.zoneList = []; 
-    }
-    
-    static deserializeBinary(bytes) {
-        const instance = new this();
-        // Simple protobuf parsing - look for repeated field 1 (zone_list)
-        let pos = 0;
-        while (pos < bytes.length) {
-            if (pos >= bytes.length) break;
-            
-            // Read varint tag
-            let tag = 0;
-            let shift = 0;
-            while (pos < bytes.length) {
-                const byte = bytes[pos++];
-                tag |= (byte & 0x7F) << shift;
-                if ((byte & 0x80) === 0) break;
-                shift += 7;
-            }
-            
-            const fieldNumber = tag >> 3;
-            const wireType = tag & 0x7;
-            
-            if (fieldNumber === 1 && wireType === 2) { // repeated zone_list field
-                // Read length
-                let length = 0;
-                shift = 0;
-                while (pos < bytes.length) {
-                    const byte = bytes[pos++];
-                    length |= (byte & 0x7F) << shift;
-                    if ((byte & 0x80) === 0) break;
-                    shift += 7;
-                }
-                
-                if (pos + length <= bytes.length) {
-                    const zoneBytes = bytes.slice(pos, pos + length);
-                    const zone = exports.ThermalZoneComponentViewModelState.deserializeBinary(zoneBytes);
-                    instance.zoneList.push(zone);
-                    pos += length;
-                } else {
-                    break;
-                }
-            } else {
-                // Skip unknown field
-                break;
-            }
-        }
-        return instance;
-    }
-    
-    getZoneListList() { 
-        return this.zoneList; 
-    }
-    
-    setZoneListList(value) { 
-        this.zoneList = value; 
-    }
-};
-
-exports.ThermalZoneComponentViewModelState = class {
-    constructor() { 
-        this.zone = 0; 
-        this.temperature = 0; 
-    }
-    
-    static deserializeBinary(bytes) {
-        const instance = new this();
-        let pos = 0;
-        
-        while (pos < bytes.length) {
-            if (pos >= bytes.length) break;
-            
-            // Read varint tag
-            let tag = 0;
-            let shift = 0;
-            while (pos < bytes.length) {
-                const byte = bytes[pos++];
-                tag |= (byte & 0x7F) << shift;
-                if ((byte & 0x80) === 0) break;
-                shift += 7;
-            }
-            
-            const fieldNumber = tag >> 3;
-            
-            if (fieldNumber === 1) { // zone field
-                let value = 0;
-                shift = 0;
-                while (pos < bytes.length) {
-                    const byte = bytes[pos++];
-                    value |= (byte & 0x7F) << shift;
-                    if ((byte & 0x80) === 0) break;
-                    shift += 7;
-                }
-                instance.zone = value;
-            } else if (fieldNumber === 2) { // temperature field
-                let value = 0;
-                shift = 0;
-                while (pos < bytes.length) {
-                    const byte = bytes[pos++];
-                    value |= (byte & 0x7F) << shift;
-                    if ((byte & 0x80) === 0) break;
-                    shift += 7;
-                }
-                instance.temperature = value;
-            } else {
-                // Skip unknown field
-                break;
-            }
-        }
-        
-        return instance;
-    }
-    
-    getZone() { 
-        return this.zone; 
-    }
-    
-    setZone(value) { 
-        this.zone = value; 
-    }
-    
-    getTemperature() { 
-        return this.temperature; 
-    }
-    
-    setTemperature(value) { 
-        this.temperature = value; 
-    }
-};
-
-// Export for CommonJS compatibility
-if (typeof module !== 'undefined' && module.exports) {
-    module.exports = exports;
-}";
-        
-        File.WriteAllText(Path.Combine(testProjectDir, "testviewmodelservice_pb.js"), protobufStubContent);
+        var baseTestDir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)!;
+        var rootDir = Path.GetFullPath(Path.Combine(baseTestDir, "../../../../.."));
+        var testDataDir = Path.Combine(rootDir, "test", "RemoteMvvmTool.Tests", "TestData", "GrpcWebEndToEnd");
+        var sourceFile = Path.Combine(testDataDir, "testviewmodelservice_pb.js");
+        File.Copy(sourceFile, Path.Combine(outDir, "testviewmodelservice_pb.js"), true);
     }
 
     static void CreateMinimalNodeModules(string nodeModulesDir)
@@ -734,26 +322,6 @@ module.exports = protobuf;");
             SetupTestProject(tempDir);
             var testProjectDir = Path.Combine(tempDir, "TestProject");
 
-            // Create the server-side test files
-            var vmCode = @"public class TestObservablePropertyAttribute : System.Attribute {}
-public class TestRelayCommandAttribute : System.Attribute {}
-namespace HP.Telemetry { public enum Zone { CPUZ_0, CPUZ_1 } }
-namespace Generated.ViewModels {
-  public class ThermalZoneComponentViewModel { public HP.Telemetry.Zone Zone { get; set; } public int Temperature { get; set; } }
-  public partial class TestViewModel : ObservableObject {
-    [TestObservableProperty]
-    private System.Collections.ObjectModel.ObservableCollection<ThermalZoneComponentViewModel> zoneList;
-  }
-}
-public class ObservableObject {}";
-            
-            File.WriteAllText(Path.Combine(testProjectDir, "TestViewModel.cs"), vmCode);
-            
-            var vmStub = @"namespace Generated.ViewModels { public partial class TestViewModel : ObservableObject { public TestViewModel() {} public System.Collections.ObjectModel.ObservableCollection<ThermalZoneComponentViewModel> ZoneList { get; set; } = new System.Collections.ObjectModel.ObservableCollection<ThermalZoneComponentViewModel>(); } }";
-            File.WriteAllText(Path.Combine(testProjectDir, "TestViewModelStub.cs"), vmStub);
-            
-            var clientStub = @"namespace Generated.Clients { public class TestViewModelRemoteClient { public TestViewModelRemoteClient(object c) {} public System.Threading.Tasks.Task InitializeRemoteAsync() => System.Threading.Tasks.Task.CompletedTask; } }";
-            File.WriteAllText(Path.Combine(testProjectDir, "TestViewModelRemoteClient.cs"), clientStub);
 
             // Generate the server code using our generators
             var refs = LoadDefaultRefs();
@@ -784,8 +352,6 @@ public class ObservableObject {}";
             var partial = ViewModelPartialGenerator.Generate(name, "Test.Protos", name + "Service", "Generated.ViewModels", "Generated.Clients", "ObservableObject", "console", true);
             File.WriteAllText(Path.Combine(testProjectDir, name + ".Remote.g.cs"), partial);
             
-            var options = @"namespace PeakSWC.Mvvm.Remote { public class ServerOptions { public int Port { get; set; } public bool UseHttps { get; set; } = false; } public class ClientOptions { public string Address { get; set; } = ""http://localhost""; } }";
-            File.WriteAllText(Path.Combine(testProjectDir, "GrpcRemoteOptions.cs"), options);
 
             // Compile and run the server
             var sourceFiles = Directory.GetFiles(testProjectDir, "*.cs").Concat(Directory.GetFiles(grpcOut, "*.cs"));
