+++ test/RemoteMvvmTool.Tests/TestData/GrpcWebEndToEnd/test-protoc.js
@@ -89,3 +89,4 @@ const process = require('process');
   console.error('Unhandled error:', e); 
   process.exit(1); 
 });
+
